+++ docs/md/releasenotes.md
@@ -20,6 +20,7 @@
 - Implement `ClassVisitor` in `KotlinMetadataInitializer` to allow easier initialization of Kotlin metadata.
 - Implement `ClassVisitor` in `KotlinMetadataWriter` to allow easier writing of Kotlin metadata.
 - Add error handler callback to `KotlinMetadataInitializer`.
+- Add error handler callback to `KotlinMetadataWriter`.
 
 ### Bug fixes